requirements.txt
beautifulsoup4==4.14.3
bs4==0.0.2
fastapi==0.135.0
fastapi-cli==0.0.24
fastapi-cloud-cli==0.14.0
python-dotenv==1.2.1
requests==2.32.5
uvicorn==0.41.0
